KafkaScheduler/requirement.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.3
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.1
requests==2.28.1
six==1.16.0
urllib3==1.26.10
Werkzeug==2.1.2
zipp==3.8.1
